fix: remove quote API usage for fiat amount and price impact calc (#3074)
* refactor: remove usd amount calc based ony CowSwap quote

* refactor: move useHigherUSDValue hook to fiatAmount module

* refactor: calculate priceImpact only basing on Coingecko

* chore: fix import

* chore: rename

* chore: remove error field from price impact

* chore: fix build

* chore: fix build
