+++ apps/cowswap-frontend/src/modules/limitOrders/containers/LimitOrdersWarnings/index.tsx
@@ -7,8 +7,6 @@ import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 import styled from 'styled-components/macro'
 import { Nullish } from 'types'
 
-import { PriceImpact } from 'legacy/hooks/usePriceImpact'
-
 import { useLimitOrdersDerivedState } from 'modules/limitOrders/hooks/useLimitOrdersDerivedState'
 import { useLimitOrdersFormState } from 'modules/limitOrders/hooks/useLimitOrdersFormState'
 import { useRateImpact } from 'modules/limitOrders/hooks/useRateImpact'
@@ -38,7 +36,6 @@ import { RateImpactWarning } from '../../pure/RateImpactWarning'
 const FORM_STATES_TO_SHOW_BUNDLE_BANNER = [TradeFormValidation.ExpertApproveAndSwap, TradeFormValidation.ApproveAndSwap]
 
 export interface LimitOrdersWarningsProps {
-  priceImpact: PriceImpact
   feeAmount?: Nullish<CurrencyAmount<Currency>>
   isConfirmScreen?: boolean
   className?: string
@@ -58,7 +55,7 @@ const StyledRateImpactWarning = styled(RateImpactWarning)`
 `
 
 export function LimitOrdersWarnings(props: LimitOrdersWarningsProps) {
-  const { priceImpact, feeAmount, isConfirmScreen = false, className } = props
+  const { feeAmount, isConfirmScreen = false, className } = props
 
   const { isPriceImpactAccepted, isRateImpactAccepted } = useAtomValue(limitOrdersWarningsAtom)
   const updateLimitOrdersWarnings = useSetAtom(updateLimitOrdersWarningsAtom)
@@ -73,8 +70,7 @@ export function LimitOrdersWarnings(props: LimitOrdersWarningsProps) {
   const tradeQuote = useTradeQuote()
 
   const canTrade = localFormValidation === null && primaryFormValidation === null && !tradeQuote.error
-  const showPriceImpactWarning =
-    canTrade && !tradeQuote.isLoading && !!chainId && !expertMode && !!account && !!priceImpact.error
+  const showPriceImpactWarning = canTrade && !tradeQuote.isLoading && !!chainId && !expertMode && !!account
   const showRateImpactWarning =
     canTrade &&
     !tradeQuote.isLoading &&
