+++ apps/cowswap-frontend/src/modules/limitOrders/containers/LimitOrdersWidget/index.tsx
@@ -225,7 +225,7 @@ const LimitOrders = React.memo((props: LimitOrdersProps) => {
           </styledEl.FooterBox>
         )}
 
-        <LimitOrdersWarnings priceImpact={priceImpact} feeAmount={feeAmount} />
+        <LimitOrdersWarnings feeAmount={feeAmount} />
 
         <styledEl.TradeButtonBox>
           <TradeButtons tradeContext={tradeContext} priceImpact={priceImpact} />
