+++ apps/cowswap-frontend/src/modules/swap/containers/TradeSummary/index.tsx
@@ -1,8 +1,8 @@
 import { Percent } from '@uniswap/sdk-core'
 
-import { useHigherUSDValue } from 'legacy/hooks/useStablecoinPrice'
 import TradeGp from 'legacy/state/swap/TradeGp'
 
+import { useHigherUSDValue } from 'modules/fiatAmount'
 import { TradeSummaryContent } from 'modules/swap/pure/TradeSummary'
 import { useWalletDetails } from 'modules/wallet'
