+++ apps/cowswap-frontend/src/modules/swap/hooks/useHandleSwap.test.tsx
@@ -42,7 +42,6 @@ const mockUseSafeBundleEthFlowContext = useSafeBundleEthFlowContext as jest.Mock
 
 const priceImpactMock: PriceImpact = {
   priceImpact: undefined,
-  error: undefined,
   loading: false,
 }
 describe('useHandleSwapCallback', () => {
