+++ apps/cowswap-frontend/src/modules/swap/state/useSwapDerivedState.ts
@@ -3,9 +3,10 @@ import { useEffect } from 'react'
 
 import { OrderKind } from '@cowprotocol/cow-sdk'
 
-import { useHigherUSDValue } from 'legacy/hooks/useStablecoinPrice'
 import { Field } from 'legacy/state/swap/actions'
 import { useDerivedSwapInfo, useSwapState } from 'legacy/state/swap/hooks'
+
+import { useHigherUSDValue } from 'modules/fiatAmount'
 
 import { useSafeMemoObject } from 'common/hooks/useSafeMemo'
