+++ apps/cowswap-frontend/src/modules/trade/containers/TradeBasicConfirmDetails/index.tsx
@@ -4,7 +4,7 @@ import { Currency, CurrencyAmount, Percent } from '@uniswap/sdk-core'
 
 import { Nullish } from 'types'
 
-import { useHigherUSDValue } from 'legacy/hooks/useStablecoinPrice'
+import { useHigherUSDValue } from 'modules/fiatAmount'
 
 import { usePrice } from 'common/hooks/usePrice'
 import { RateInfoParams } from 'common/pure/RateInfo'
