+++ apps/cowswap-frontend/src/modules/trade/state/priceImpactAtom.ts
@@ -4,6 +4,5 @@ import { PriceImpact } from 'legacy/hooks/usePriceImpact'
 
 export const priceImpactAtom = atom<PriceImpact>({
   priceImpact: undefined,
-  error: undefined,
   loading: false,
 })
