+++ apps/cowswap-frontend/src/modules/trade/updaters/PriceImpactUpdater.tsx
@@ -1,20 +1,20 @@
 import { useSetAtom } from 'jotai'
 
-import { usePriceImpact } from 'legacy/hooks/usePriceImpact'
+import { useFiatValuePriceImpact } from 'legacy/hooks/usePriceImpact'
 
 import { useSafeEffect } from 'common/hooks/useSafeMemo'
 
-import { useTradePriceImpactParams } from '../hooks/useTradePriceImpactParams'
+import { useTradeParsedAmounts } from '../hooks/useTradeParsedAmounts'
 import { priceImpactAtom } from '../state/priceImpactAtom'
 
 export function PriceImpactUpdater() {
   const updatePriceImpact = useSetAtom(priceImpactAtom)
-  const params = useTradePriceImpactParams()
-  const { error, loading, priceImpact } = usePriceImpact(params)
+  const params = useTradeParsedAmounts()
+  const { isLoading, priceImpact } = useFiatValuePriceImpact(params)
 
   useSafeEffect(() => {
-    updatePriceImpact({ error, loading, priceImpact })
-  }, [error, loading, updatePriceImpact, priceImpact])
+    updatePriceImpact({ loading: isLoading, priceImpact })
+  }, [isLoading, updatePriceImpact, priceImpact])
 
   return null
 }
