+++ apps/cowswap-frontend/src/modules/tradeQuote/hooks/useTradeQuoteFeeFiatAmount.ts
@@ -3,7 +3,7 @@ import { useMemo } from 'react'
 
 import { CurrencyAmount, Token } from '@uniswap/sdk-core'
 
-import { useHigherUSDValue } from 'legacy/hooks/useStablecoinPrice'
+import { useHigherUSDValue } from 'modules/fiatAmount'
 
 import { useTradeQuote } from './useTradeQuote'
