+++ apps/cowswap-frontend/src/modules/twap/hooks/useTwapWarningsContext.ts
@@ -1,6 +1,5 @@
 import { useMemo } from 'react'
 
-import { useTradePriceImpact } from 'modules/trade'
 import { TradeFormValidation, useGetTradeFormValidation } from 'modules/tradeFormValidation'
 import { useWalletInfo } from 'modules/wallet'
 
@@ -19,13 +18,12 @@ export interface TwapWarningsContext {
 export function useTwapWarningsContext(): TwapWarningsContext {
   const { account } = useWalletInfo()
   const primaryFormValidation = useGetTradeFormValidation()
-  const priceImpact = useTradePriceImpact()
 
   return useMemo(() => {
     // TODO: bind to settings
     const expertMode = false
     const canTrade = !primaryFormValidation || NOT_BLOCKING_VALIDATIONS.includes(primaryFormValidation)
-    const showPriceImpactWarning = canTrade && !expertMode && !!priceImpact.error
+    const showPriceImpactWarning = canTrade && !expertMode
     const walletIsNotConnected = !account
 
     return {
@@ -33,5 +31,5 @@ export function useTwapWarningsContext(): TwapWarningsContext {
       showPriceImpactWarning,
       walletIsNotConnected,
     }
-  }, [priceImpact, primaryFormValidation, account])
+  }, [primaryFormValidation, account])
 }
